requirements.txt
affine==2.2.1
alabaster==0.7.11
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
base==1.0.4
certifi==2018.8.24
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.4.0
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
docutils==0.14
GDAL==2.2.3
idna==2.7
imagesize==1.1.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
networkx==2.2
numpy==1.14.5
numpydoc==0.8.0
olefile==0.46
packaging==17.1
pandas==0.23.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.1.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
props==0.0.3
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.1
pyproj==1.9.5.1
pyshp==1.2.12
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
pyzmq==17.1.2
rasterio==1.0.6
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
snuggs==1.4.2
Sphinx==1.8.0
sphinxcontrib-websupport==1.1.0
toolz==0.9.0
torch==0.4.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
